less shitcode!! yay!
