+++ Resources/Prototypes/Entities/Objects/Tools/lighters.yml
@@ -12,6 +12,8 @@
       collection: lighterOnSounds
     soundDeactivate:
       collection: lighterOffSounds
+    soundFailToActivate:
+      collection: lighterFailSounds
   - type: ItemToggleMeleeWeapon
     activatedDamage:
       types:
@@ -124,8 +126,6 @@
           cheap_icon_base: Rainbow
   - type: CheapLighter
     failChance: 0.4
-    soundFail:
-      collection: lighterFailSounds
   - type: Welder
     fuelConsumption: 0.02
     fuelLitCost: 0.2
@@ -148,8 +148,6 @@
       map: [ "flame" ]
   - type: CheapLighter
     failChance: 0.5
-    soundFail:
-      collection: lighterFailSounds
   - type: Welder # the single worst lighter invented by mankind
     fuelConsumption: 0.4
     fuelLitCost: 0.2
